backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pymysql==1.1.0
cryptography==42.0.5
bcrypt==4.1.2
pydantic[email]==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9
